ai_agent_service/requirements.txt
fastapi>=0.111.0
uvicorn>=0.29.0
pydantic==2.8.2
langchain==0.2.16
langchain-core==0.2.43
langchain-community==0.2.16
langchain-google-genai==1.0.8
requests>=2.31.0
python-dotenv>=1.0.1
pymongo>=4.6.0
langchain-mongodb>=0.1.0

